+++ SDP/IML/IML/Functions.h
@@ -4,20 +4,20 @@
 class Functions
 {
 public:
-	void increase(std::list<int>&, int value);
-	void multiply(std::list<int>&, int value);
+	void increase(std::list<double>&, double value);
+	void multiply(std::list<double>&, double value);
 
-	void agg_sum(std::list<int>&);
-	void agg_pro(std::list<int>&);
-	void agg_avg(std::list<int>&);
-	void agg_first(std::list<int>&);
-	void agg_last(std::list<int>&);
+	void agg_sum(std::list<double>&);
+	void agg_pro(std::list<double>&);
+	void agg_avg(std::list<double>&);
+	void agg_first(std::list<double>&);
+	void agg_last(std::list<double>&);
 
-	void reverse(std::list<int>&);
-	void sort_asc(std::list<int>&);
-	void sort_dsc(std::list<int>&);
-	void slice(std::list<int>&, size_t position);
-	void remove_duplicates(std::list<int>&);
+	void reverse(std::list<double>&);
+	void sort_asc(std::list<double>&);
+	void sort_dsc(std::list<double>&);
+	void slice(std::list<double>&, size_t position);
+	void remove_duplicates(std::list<double>&);
 };
